% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistanceFilter.R
\name{DistanceFilter}
\alias{DistanceFilter}
\title{DistanceFilter - Raryfy the points in the dataset with threshold distance}
\usage{
DistanceFilter(input = NA, threshold = 1, output = NA)
}
\arguments{
\item{input}{- input file name containing species name, Latitude and Longitude}

\item{threshold}{- value of threshold in (decimal) degrees}

\item{output}{- name of the output file}
}
\description{
DistanceFilter - Raryfy the points in the dataset with threshold distance
}
\examples{
\dontrun{
DistanceFilter()
}
}
