% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBatchMask.R
\name{iBatchMask}
\alias{iBatchMask}
\title{iBatchMask - Crops (masks) the raster files with multiple shape file. (Interactive)}
\usage{
iBatchMask(ASCfilelist = NA, SHPfilelist = NA, OPDirName = NA)
}
\arguments{
\item{ASCfilelist}{- list of raster files to crop}

\item{SHPfilelist}{- mask shapefiles used to crop rasters}

\item{OPDirName}{- Output folder, where subfolders of shape files will be created}
}
\description{
Interactive version. For interactive version refer \link{BatchMask}
}
\details{
Function crops (masks) the raster files with multiple shape file and stores 
them in separate folders. Interactive varions. For noninteractive version to
use in scripts check \link{iBatchMask}

Masked files are stored in seperate folders with shape file name and names of the masked files are kept same. 
For example, if the input asc files are bio1.asc, bio2.asc, and you select multiple shape files like shp1.shp, shp2.shp, shp3.shp, then 
3 folders named shp1, shp2 and shp3 will be created in output folder specified by user. Each folder will contain bio1.asc, bio2.asc
If by chance your files does not have any extension, then comment the statement below the note 'For file with extension'. 
and uncomment the FileName statement below the note 'For files without extension
}
\examples{
\dontrun{
iBatchMask()
}
}
