% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCropRaster.R
\name{iCropRaster}
\alias{iCropRaster}
\title{iCropRaster - Crops (masks) the raster files with the shape file. (Interactive)}
\usage{
iCropRaster(filelist = NA, ShapeFile = NA, sufix = NA)
}
\arguments{
\item{filelist}{- list of raster files to crop}

\item{ShapeFile}{- mask shapefile used to crop rasters}

\item{sufix}{- sufix for the output file name}
}
\description{
Function crops (masks) the raster files with the shape file. Interactive 
version. For noninteractive veriosn to be used in scripts check \link{CropRaster}
}
\details{
Masked files are stored in the current working directory with original name of 
file + the sufix provided.
For example, if the input asc files are bio1.asc, bio2.asc, and you give the 
suffix as 'n', then the output masked files are bio1n.asc, bio2n.asc
If by chance your files do not have any extension, then comment the statement 
below the note 'For file with extension' and uncomment the FileName statement 
below the note 'For files without extension
}
\examples{
\dontrun{
iCropRaster()
}
}
