% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMOP.R
\name{iMOP}
\alias{iMOP}
\title{iMOP - This function calculates distance measures between numerous groups of reference set of points (M) in n-dimensional space with each point in another set (G).}
\usage{
iMOP(FileType = NA, m1 = NA, m2 = NA, InpRefFile = NA,
  InpExtentFile = NA, p1 = NA, p2 = NA, decil = NA, MxMESS = NA,
  GetAns = NA, OPRefFileName = NA, OPExtentFileName = NA, FlList = NA,
  SampleMName = NA, SampleGName = NA, MxMESSOpFile = NA)
}
\arguments{
\item{FileType}{- File types of reference and extent area. 
A - ASC format files. Parameter m1 and m2 are required, if FileType is set to A 
M - Files in text format, for files in text format. Parameter InpRefFile and InpExtentFile are required, if FileType is set to M
The format of the file should be Longitude, Latitude, Var1, Var2,...... and so on. Also make sure that reference and extent 
file in text format has same sequence of variables.}

\item{m1}{- List of .asc file for reference area (M). For example, ("C://data//Bio1_M.asc", "C://data//Bio2_M.asc","C://data//Bio3_M.asc")
Supply this list, when FileType parameter is set to A. Make sure that order of variables is same in reference and exent.}

\item{m2}{- List of .asc file for extent area (G). For example, ("C://data//Bio1_G.asc", "C://data//Bio2_G.asc","C://data//Bio3_G.asc"). 
Supply this list, when FileType parameter is set to A. Make sure that order of variables is same in reference and exent}

\item{InpRefFile}{- Name of file containing reference data (M) for distance calculation. Format of the file should be 
Longitude, Latitude, Var1, Var2,...... and so on.}

\item{InpExtentFile}{- Name of file containing extent data (G) for distance calculation. Format of the file should be 
Longitude, Latitude, Var1, Var2,...... and so on. Order of variables in Reference (M) and Extent (G) data should be maintained.}

\item{p1}{- Subsampling percentage for reference area (M) (proportion of points from M to be randomly sampled if M is too large). 
Value must be between (>0 and <=1).}

\item{p2}{- Subsampling percentage for Extent area (G) (proportion of points from G to be randomly sampled if G is too large). 
Value must be between (>0 and <=1).}

\item{decil}{- Series of number seperated by comma, which signifies what proportion of closest points in M is to be compared.}

\item{MxMESS}{- A boolean variable for generating Maxent MESS, if MxMESS = Y, generate Maxent MESS , if MxMESS = N do not generate Maxent MESS}

\item{GetAns}{- Option is Y or N. If Y, the provide file name to save reference data and extent data seperately. This parameter is necessary, 
if the FileType parameter is set to A.}

\item{OPRefFileName}{- Name of file in which reference data (M) will be saved in .txt format for future use. Supply this parameter when
GetAns = Y and FileType = A}

\item{OPExtentFileName}{- Name of file in which extent data (G) will be saved in .txt format for future use. Supply this parameter when
GetAns = Y and FileType = A}

\item{FlList}{- List for filenames with folder names. This files will store the generated MOP maps. Number of filenames should be equal to 
number of decil provided}

\item{SampleMName}{- Provide file name with folder names. If the reference area is too big and p1 is less than 1, then choice is to save this resampled data in the file for 
future use.}

\item{SampleGName}{- Provide file name with folder names. If the reference area is too big and p1 is less than 1, then choice is to save this resampled data in the file for 
future use.}

\item{MxMESSOpFile}{- If MxMESS is set to Y, and want to save maxent MESS results in a file, then provide file name with folder in this parameter.}
}
\description{
Interactive version. For noninteractive version refer \link{MOP}
}
\details{
Function MOP also marks areas where extrapolation could occur during niche modeling exercise for careful model interpretation. 
This function is especially useful when developed niche model in projected to different space or time. 
For more details refer DOI: 10.1016/j.ecolmodel.2013.04.011
The function will return a list of length 3 or 4, depending upon MxMESS parameter. If MxMESS is set to Y then it returns a list of 3 items
First object in list is - MOP matrix, Second object in list is ? Maxent MESS matrix, Third object in list is - Matrix of resampled M. 
If MxMESS parameter is to N then return list contains 2 objects, 1st - MOP matrix, 2nd - Matrix of resampled M
Input parameters required for this function are
}
\examples{
\dontrun{
iMOP()
}
}

