% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPartialROC.R
\name{iPartialROC}
\alias{iPartialROC}
\title{iPartialROC - A function to evaluate output of presence only nich models on continuous model outputs.}
\usage{
iPartialROC(PresenceFile = NA, PredictionFile = NA, OmissionVal = NA,
  RandomPercent = NA, NoOfIteration = NA, OutputFile = NA)
}
\arguments{
\item{PresenceFile}{- occurence point file (test data)}

\item{PredictionFile}{- model prediction in .asc format}

\item{OmissionVal}{- Amount of error admissible along the Y-axis, given the requirements and conditions of the study. Value should range between 0 - 1}

\item{RandomPercent}{- Occurrence points to be sampled randomly from the test data for bootstrapping.}

\item{NoOfIteration}{- Number of iteration for bootstrapping}

\item{OutputFile}{- Name of output file. If you specify the path, it will store it there otherwise it will save it in current working directory.}
}
\description{
Interactive version. For noninteractive version refer \link{PartialROC}
}
\details{
Function PartialROC generates the area under the curve values using bootstrap method. PartialROC is a model evaluation tool, used for
continuous model outputs as compared to binary model outputs. This method is specially used for model trained using presence only data.
For more details refer DOI: 10.1016/j.ecolmodel.2007.11.008
Input parameters required for this function are 
   1. PresenceFile - (test data set) in .csv format, must have 3 columns SpName, Longitude, Latitude.
   2. PredictionFile - (Thresholded continuous model output). This file should be in .asc format
   3. OmissionVal - Amount of error admissible along the Y-axis, given the requirements and conditions of the study. Value should range
                    between 0 - 1
   4. RandomPercent - Occurrence points to be sampled randomly from the test data for bootstrapping.
   5. NoOfIteration - Number of iteration for bootstrapping
   6. OutputFile - Name of output file. If you specify the path, it will store it there otherwise it will save it in current working
                   directory. 
OutputFile will have 4 columns, IterationNo, AUC_at_specified_value, AUC_AT_Random, AUC_Ratio. The first row will always have 0 th interation
which is the actual Area Under the Curve without bootstrapping. And the rest of the rows contains auc ratio for all the bootstrap.
}
\examples{
\dontrun{
iPartialROC()
}
}
