% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPCAProjection.R
\name{iPCAProjection}
\alias{iPCAProjection}
\title{iPCAProjection - PCA of Raster files (Interactive)}
\usage{
iPCAProjection(BioStackFiles = NA, LoadingFile = NA, CompImpFile = NA,
  ProjectonStackFiles = NA, OutputFolder = NA)
}
\arguments{
\item{BioStackFiles}{- ESRI ASCII grid files of predictor variables}

\item{LoadingFile}{- name of output file which stores loadings for components}

\item{CompImpFile}{- name of output file of the PCA summaries}

\item{ProjectonStackFiles}{- Future climate ESRI ASCII grid files or a list of lists for multiple projections}

\item{OutputFolder}{- Output folder or list of output folder names for storing the projection(s)}
}
\value{
a summary of the PCA is returnd as a strcture
}
\description{
Performs PCSRaster and projects it on another set of rasters.
}
\details{
Interactive version. For noninteractive veriosn to be used in scripts refer \link{PCAProjection}
}
\examples{
\dontrun{
pcaop = iPCAProjection()
}
}

