% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelThreshold.r
\name{ModelThreshold}
\alias{ModelThreshold}
\title{ModelThreshold - Thresholds Maxent suitability prediction based on user input. Output of the function is binary map of
predicted suitability as 1 and non suitabile region as 0.}
\usage{
ModelThreshold(InSuitFile = NA, OccurrenceFile = NA,
  PercentThreshold = NA, OutSuitFile = NA)
}
\arguments{
\item{InSuitFile}{- Maxent prediction in .asc format}

\item{OccurrenceFile}{- Occurrences using which Maxent model is trained.}

\item{PercentThreshold}{- Percent of ommission error, ranges between 0 - 1}

\item{OutSuitFile}{- Output file name. This file will be stored in the current working directory, if no path is given.}
}
\description{
Noninteractive version. For interacteve version refer \link{iModelThreshold}
}
\details{
Function thresholds the suitability predicted by Maxent or by any other niche modeling algorithm, provided that 
predicted file is in .asc format. Function needs 4 parameters, input file, which is prediction file. Occurrence file, 
this a occurrence file used in calibrating the model. Format of this file is, speciesname, Longitude, Latitude.
PercentThreshold - what percent of ommission is agreable while calibrating the model. This percent is suppose to be decided
depending upon how the occurrences are sampled. 
OutSuitFile - Final thresholded prediction will be stored in this file and the format of this file is .asc. Suitability
will be reclassified into 0 and 1. All the pixels above the thresholded value will be classified as 1 and the pixels below
threshold value will be classified as 0.
}
\examples{
\dontrun{
ModelThreshold()
}
}
