% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheViews.R
\name{NicheViews}
\alias{NicheViews}
\title{NicheViews -}
\usage{
NicheViews(OutputSufix = NA, RasterFileList = NA, DataFileName = NA)
}
\arguments{
\item{OutputSufix}{- Suffix for the output graph files. (file names will be 
"sample12.jpg" where sample is suffix and component 1 and 2 are plotted in this 
plot)}

\item{RasterFileList}{- Component files in .asc format (Use the components 
generated by PCA function, you can use any .asc files if you need)}

\item{DataFileName}{- Species occurrences file - .csv file with "Species, 
Longitude, Latitude" fields with headers.}
}
\description{
Function to generate bivariate graphs of all combinations of PCA components, 
overlaid with occurrence data. Graphs are stored in jpg files. This 
function also generates bar plots of each component, occurence of that components.
}
\details{
Nininteractive version.

Time required to run the function depends upon the size of component files.
}
\examples{
\dontrun{
NicheViews()
}
}
