% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAProjection.R
\name{PCAProjection}
\alias{PCAProjection}
\title{PCAProjection - PCA of Raster files}
\usage{
PCAProjection(BioStackFiles = NA, LoadingFile = NA, CompImpFile = NA,
  ProjectonStackFiles = NA, OutputFolder = NA)
}
\arguments{
\item{BioStackFiles}{- ESRI ASCII grid files of predictor variables}

\item{LoadingFile}{- name of output file which stores loadings for components}

\item{CompImpFile}{- name of output file of the PCA summaries}

\item{ProjectonStackFiles}{- Future climate ESRI ASCII grid files or a list of 
lists for multiple projections}

\item{OutputFolder}{- Output folder or list of output folder names for storing 
the projection(s)}
}
\value{
a summary of the PCA is returnd as a strcture
}
\description{
Performs PCSRaster and projects it on another set of rasters
}
\details{
Noninteractive version. For interactive veriosn to be used in scripts refer \link{iPCAProjection}
}
\examples{
\dontrun{
pcaop = PCAProjection()
}
}

