% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCARaster.R
\name{PCARaster}
\alias{PCARaster}
\title{PCARaster - PCA of Raster files}
\usage{
PCARaster(BioStackFiles = NA, LoadingFile = NA, CompImpFile = NA,
  OPfolder = NA)
}
\arguments{
\item{BioStackFiles}{- ESRI ASCII grid files of predictor variables}

\item{LoadingFile}{- name of output file which stores loadings for components}

\item{CompImpFile}{- name of output file of the PCA summaries}

\item{OPfolder}{- name of output folder to save PCA component, loading and summary file}
}
\value{
a summary of the PCA is returnd as a strcture
}
\description{
Performs Principle Component Analysis of Raster objects and returns summary 
and loadings for the componetns. For interactive version check \link{iPCARaster}
}
\details{
Main function to generate PCA for the selected bioclimatic layer and then save 
the pca components in ASCII format.
This function will accept the bioclimatic ASCII files. PCApca components are 
stored as Comp1.asc, Comp2.asc...... and so on. 
This function retuns the pca of the ascii data supplied to do further processing like checking for eigen values, broken stick etc.
}
\examples{
\dontrun{
pcaop = PCARaster()
}
}

